+++ project/components/modals/delete-project-modal.tsx
@@ -2,8 +2,8 @@ import { deleteProject } from "@/lib/actions/projects";
 import { ProjectSchema } from "@/types/projects";
 import { redirect, RedirectType } from "next/navigation";
 import { ChangeEvent, useState } from "react";
-import Button from "../button";
 import Input from "../input";
+import Button from "../ui/buttons/button";
 import { showErrorToast, showSuccessToast } from "../ui/toast";
 import Modal from "./modal";
