+++ project/components/settings-navigation.tsx
@@ -1,6 +1,6 @@
 import { cn } from "@/lib/utils";
 import { Bell, Palette, Shield, User } from "lucide-react";
-import Button from "./button";
+import Button from "./ui/buttons/button";
 
 type SettingsItem = {
   name: string;
